+++ angular-workspace/components/info-button/info-button.stories.ts
@@ -19,7 +19,7 @@ const meta: Meta<InfoButtonArgs> = {
 
 export default meta;
 
-const { Default, Information, SecondaryActive, SecondaryInactive } = infoButtonStories({
+const { Default, Information } = infoButtonStories({
   templateContainer,
   storyTemplates: (templates) => {
     const { infoButtonTemplate } = templates;
@@ -31,4 +31,4 @@ const { Default, Information, SecondaryActive, SecondaryInactive } = infoButtonS
   },
 });
 
-export { Default, Information, SecondaryActive, SecondaryInactive };
+export { Default, Information };